+++ CHANGELOG.md
@@ -389,6 +389,10 @@ Released with 1.0.0-beta.37 code base.
 3. `defaultAccount` default value is undefined
 4. `defaultNetworkId` default value is undefined
 
+##### web3-eth-subscribe
+
+1. `clearSubscriptions` Instead of returning `true` ,  `clearSubscriptions` now returns array of subscription's ids
+
 #### web3-eth-personal
 
 1. `givenProvider` default value is undefined